Move db initializing in separate file Added db_version table
